Create README.md
I am a  second year PhD student of C. K. Tedam University of Technology and Applied Science in Navrongo, Ghana. My thesis topic is on how to use homomorphic and multi-party encryption to protect e-health data or records on the cloud. I will need the assistance of all to enable me achieve success. Thanks for your support in advance.